The CommonJS / ESM feud has gone too far.
