+++ package.json
@@ -1,6 +1,5 @@
 {
   "name": "hazel",
-  "type": "module",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/hazelgrove/hazel.git"
